Lost track of all the - many - changes.
Added some 3D planar geometry primitives, including line and plane
objects as well as functions for intersection, parallelism testing,
distance testing, etc.  It may make sense to move the planar operations
into a separate header.  Currently, the geometry has some basic testing
in the unit tests.  It would be good to add more, however, as it is
clearly a complicated portion of the library.

Finally implemented the Kalman filtering of tracker position/velocity
estimation and splashdown prediction point in the
gnss_tracker_protocol.h file.  These are being tested as part of the
splashdown estimation test program (in Chandra-Rocketry), but they are not actually being
tested directly in the unit testing harness.  The KF implementation did
lead to the discovery of a rather annoying bug in the base matrix
library, which has been fixed.  It was not correctly accessing values in
row vectors when using two indicies.

Otherwise, there are many other small revisions in other areas.

+++ Chandra HAL/circular_buffer.h
@@ -154,7 +154,7 @@ class NonblockingFixedCircularBuffer
           return buffer_[idx];
         }
 
-        ref_t drop(uint8_t _num) {
+        ref_t drop(size_t _num) {
         	_num = (_num>cnt_) ? cnt_ : _num;
         	idx_ += _num;
         	cnt_ -= _num;
